Refactor user condition types to extend AbstractUserIsNullConditionType and update filter logic
