+++ wcfsetup/install/files/lib/system/condition/type/user/UserIsBannedConditionType.class.php
@@ -2,8 +2,6 @@
 
 namespace wcf\system\condition\type\user;
 
-use wcf\data\DatabaseObjectList;
-
 /**
  * @author Olaf Braun
  * @copyright 2001-2025 WoltLab GmbH
@@ -15,25 +13,5 @@ final class UserIsBannedConditionType extends AbstractUserBooleanConditionType
     public function __construct()
     {
         parent::__construct("isBanned", 'banned');
-    }
-
-    #[\Override]
-    public function applyFilter(DatabaseObjectList $objectList): void
-    {
-        if ($this->filter) {
-            $objectList->getConditionBuilder()->add("{$objectList->getDatabaseTableAlias()}.banned = ?", [1]);
-        } else {
-            $objectList->getConditionBuilder()->add("{$objectList->getDatabaseTableAlias()}.banned = ?", [0]);
-        }
-    }
-
-    #[\Override]
-    public function matches(object $object): bool
-    {
-        if ($this->filter) {
-            return (bool)$object->banned;
-        } else {
-            return !$object->banned;
-        }
     }
 }
