+++ wcfsetup/install/files/lib/system/condition/type/user/UserIsEmailConfirmedConditionType.class.php
@@ -8,7 +8,7 @@ namespace wcf\system\condition\type\user;
  * @license GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
  * @since 6.3
  */
-final class UserIsEmailConfirmedConditionType extends AbstractUserBooleanConditionType
+final class UserIsEmailConfirmedConditionType extends AbstractUserIsNullConditionType
 {
     public function __construct()
     {
